Moved subscription providers to dynamic chunks. Closes #1393
