+++ packages/altair-app/src/app/services/subscriptions/subscription-provider-registry.service.ts
@@ -1,15 +1,15 @@
 import { Injectable } from '@angular/core';
 import { SubscriptionProviderConstructor } from './subscription-provider';
-import { WebsocketSubscriptionProvider, WEBSOCKET_PROVIDER_ID } from './providers/ws';
-import { AppSyncSubscriptionProvider, APP_SYNC_PROVIDER_ID } from './providers/app-sync';
-import { ActionCableSubscriptionProvider, ACTION_CABLE_PROVIDER_ID } from './providers/action-cable';
 
 interface SubscriptionProviderData {
   id: string;
-  providerClass: SubscriptionProviderConstructor;
+  getProviderClass: () => Promise<SubscriptionProviderConstructor>;
   copyTag?: string;
 }
 
+export const WEBSOCKET_PROVIDER_ID = 'websocket';
+export const APP_SYNC_PROVIDER_ID = 'app-sync';
+export const ACTION_CABLE_PROVIDER_ID = 'action-cable';
 @Injectable()
 export class SubscriptionProviderRegistryService {
   private list: SubscriptionProviderData[] = [];
@@ -17,19 +17,19 @@ export class SubscriptionProviderRegistryService {
   constructor() {
     this.addProviderData({
       id: WEBSOCKET_PROVIDER_ID,
-      providerClass: WebsocketSubscriptionProvider,
+      getProviderClass: async() => (await import('./providers/ws')).WebsocketSubscriptionProvider,
       copyTag: 'SUBSCRIPTION_PROVIDER_WEBSOCKET',
     });
 
     this.addProviderData({
       id: APP_SYNC_PROVIDER_ID,
-      providerClass: AppSyncSubscriptionProvider,
+      getProviderClass: async() => (await import('./providers/app-sync')).AppSyncSubscriptionProvider,
       copyTag: 'SUBSCRIPTION_PROVIDER_APP_SYNC',
     });
 
     this.addProviderData({
       id: ACTION_CABLE_PROVIDER_ID,
-      providerClass: ActionCableSubscriptionProvider,
+      getProviderClass: async() => (await import('./providers/action-cable')).ActionCableSubscriptionProvider,
       copyTag: 'SUBSCRIPTION_PROVIDER_ACTION_CABLE',
     });
   }
